# Use Arc<str> for path component keys