--- a/crates/jsonmodem/Cargo.toml
+++ b/crates/jsonmodem/Cargo.toml
@@ -12,24 +12,20 @@ rust-version = "1.85"
 [features]
 default = []
 fuzzing = []
-serde = []
+serde = ["dep:serde"]
 bench = []
 comparison = []
-pyo3 = ["dep:pyo3"]
 bench-fast = []
 test-fast = []
 # Enabling `miri` pulls in the faster test and benchmark configurations.
 miri = ["bench-fast", "test-fast"]
 
 [dependencies]
-ouroboros = { version = "0.18.5", default-features = false }
-serde = { version = "1.0", features = ["derive"] }
-pyo3 = { version = "0.25.1", optional = true }
-cfg-if = "1.0.1"
+serde = { version = "1.0", optional = true, features = ["derive", "rc"] }
 
 [dev-dependencies]
 insta = { version = "1.43.1", features = ["yaml"] }
-serde = { version = "1.0", features = ["derive"] }
+serde = { version = "1.0", features = ["derive", "rc"] }
 quickcheck = "1.0"
 rstest = "0.25.0"
 quickcheck_macros = "1.1.0"