This improves the `streaming_json_large/streaming_parser_none/5000` benchmark by another ~10% with a smaller improvement to `streaming_json_large/streaming_values_parser/5000` (full partial JSON value iterator.)

For the latter, we will likely optimize in Rust via persistent data structures as in https://github.com/AaronFriel/jsonmodem/pull/56.

Removes unnecessary dependencies with default features.

Before:
```
streaming_json_large/streaming_parser_none/5000
                        time:   [668.03 µs 670.06 µs 672.45 µs]
streaming_json_large/streaming_parser_roots/5000
                        time:   [746.90 µs 749.92 µs 753.45 µs]
streaming_json_large/streaming_parser_all/5000
                        time:   [815.34 µs 817.98 µs 820.89 µs]
streaming_json_large/streaming_values_parser/5000
                        time:   [23.458 ms 23.523 ms 23.597 ms]
```

After:

```
streaming_json_large/streaming_parser_none/5000
                        time:   [605.09 µs 605.44 µs 605.88 µs]
streaming_json_large/streaming_parser_roots/5000
                        time:   [637.20 µs 638.47 µs 640.72 µs]
streaming_json_large/streaming_parser_all/5000
                        time:   [677.24 µs 677.53 µs 677.84 µs]
streaming_json_large/streaming_values_parser/5000
                        time:   [14.294 ms 14.358 ms 14.420 ms]
```

Fairly significant improvement across the board.

------
https://chatgpt.com/codex/tasks/task_e_68992414aba48320ab69880e72e8af89